style(color-picker): remove d-advanced-color and organize module imports
