+++ devui/color-picker/advanced-color-panel/advanced-color-panel.component.html
