+++ devui/color-picker/advanced-color-slider/advanced-color-slider.component.scss
