--- devui/color-picker/advanced-color/advanced-color.component.html
@@ -1,4 +1,0 @@
-<div class="advanced-color">
-  <d-advanced-color-panel></d-advanced-color-panel>
-  <d-color-slider></d-color-slider>
-</div>
