--- devui/color-picker/advanced-color/advanced-color.component.scss
@@ -1,5 +1,0 @@
-.advanced-color {
-    width: 100%;
-    height: 130px;
-    user-select: none;
-}
